Extract TableRow for mobile
